# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IUser": {
				"properties": {
					"key": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"user": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string",
						"nullable": true
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"key",
					"email",
					"user",
					"name",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EOrder": {
				"enum": [
					"name",
					"created",
					"type"
				],
				"type": "string"
			},
			"Integer": {
				"type": "integer",
				"format": "int32"
			},
			"IUserFilters": {
				"properties": {
					"order": {
						"$ref": "#/components/schemas/EOrder"
					},
					"name": {
						"type": "string"
					},
					"rolId": {
						"type": "string"
					},
					"count": {
						"$ref": "#/components/schemas/Integer"
					},
					"page": {
						"$ref": "#/components/schemas/Integer"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUser.Exclude_keyofIUser.password__": {
				"properties": {
					"key": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"user": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				},
				"required": [
					"key",
					"email",
					"user",
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"IUserResponse": {
				"properties": {
					"key": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"user": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				},
				"required": [
					"key",
					"email",
					"user",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUser.email-or-user-or-name-or-lastName-or-password_": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"user": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email",
					"user",
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"IUserCreationParams": {
				"$ref": "#/components/schemas/Pick_IUser.email-or-user-or-name-or-lastName-or-password_"
			},
			"Pick_IUser.email-or-user-or-name-or-lastName_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"user": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"user",
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"IUserUpdateParams": {
				"$ref": "#/components/schemas/Pick_IUser.email-or-user-or-name-or-lastName_"
			},
			"Pick_IUser.email-or-password_": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"IUserLogin": {
				"$ref": "#/components/schemas/Pick_IUser.email-or-password_"
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "apiKey",
				"name": "Authorization",
				"scheme": "bearer",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "api-nolatech",
		"version": "1.0.0",
		"description": "prueba tecnica nolatech",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/v1/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "boolean"
										},
										"data": {
											"properties": {
												"message": {
													"anyOf": [
														{
															"type": "string"
														},
														{}
													]
												},
												"users": {
													"items": {
														"$ref": "#/components/schemas/IUser"
													},
													"type": "array"
												}
											},
											"required": [
												"users"
											],
											"type": "object"
										}
									},
									"required": [
										"status",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Obtener todos los usuarios con paginación.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/EOrder"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "rolId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "count",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Integer"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Integer"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"summary": "Crear un nuevo usuario.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "- Datos del nuevo usuario.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserCreationParams",
								"description": "- Datos del nuevo usuario."
							}
						}
					}
				}
			}
		},
		"/v1/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"allOf": [
												{
													"$ref": "#/components/schemas/IUserResponse"
												}
											],
											"nullable": true
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Obtener un usuario por ID.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"description": "- ID del usuario.",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "boolean"
										},
										"data": {
											"allOf": [
												{
													"$ref": "#/components/schemas/IUserResponse"
												}
											],
											"nullable": true
										}
									},
									"required": [
										"status",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Actualizar un usuario por ID.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"description": "- ID del usuario.",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- Datos actualizados del usuario.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdateParams",
								"description": "- Datos actualizados del usuario."
							}
						}
					}
				}
			}
		},
		"/v1/users/{key}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "boolean"
										}
									},
									"required": [
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar un usuario por ID.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"description": "- ID del usuario.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "boolean"
										},
										"data": {
											"properties": {
												"message": {
													"anyOf": [
														{
															"type": "string"
														},
														{
															"items": {
																"type": "string"
															},
															"type": "array"
														}
													]
												},
												"token": {
													"type": "string"
												},
												"user": {
													"allOf": [
														{
															"$ref": "#/components/schemas/IUserResponse"
														}
													],
													"nullable": true
												}
											},
											"required": [
												"user"
											],
											"type": "object"
										}
									},
									"required": [
										"status",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Registrar un nuevo usuario.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "- Parámetros de creación del usuario.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserCreationParams",
								"description": "- Parámetros de creación del usuario."
							}
						}
					}
				}
			}
		},
		"/v1/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Usuario autenticado y token de autenticación o mensaje de error.\r\n{ data: { user: IUserResponse | null; token?: string, message?: string }, status: boolean }",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "boolean"
										},
										"data": {
											"properties": {
												"message": {
													"type": "string"
												},
												"token": {
													"type": "string"
												},
												"user": {
													"allOf": [
														{
															"$ref": "#/components/schemas/IUserResponse"
														}
													],
													"nullable": true
												}
											},
											"required": [
												"user"
											],
											"type": "object"
										}
									},
									"required": [
										"status",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Inicia sesión con credenciales de usuario.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "- Credenciales de inicio de sesión (email o user y contraseña).",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserLogin",
								"description": "- Credenciales de inicio de sesión (email o user y contraseña)."
							}
						}
					}
				}
			}
		},
		"/v1/auth/me": {
			"get": {
				"operationId": "Me",
				"responses": {
					"201": {
						"description": "User Found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "boolean"
										},
										"data": {
											"properties": {
												"user": {
													"allOf": [
														{
															"$ref": "#/components/schemas/IUserResponse"
														}
													],
													"nullable": true
												}
											},
											"required": [
												"user"
											],
											"type": "object"
										}
									},
									"required": [
										"status",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Obtiene los detalles del usuario autenticado.",
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}